requirements_deployment.txt
decorator==4.0.11
diff-match-patch==20121119
distributed==1.16.3
Django==1.11.3
django-admin-import==0.3.0
django-autoslug==1.9.3
django-bootstrap3==9.0.0
django-braces==1.11.0
django-import-export==0.6.0
django-mailbox==4.6.0
djangorestframework==3.6.4
django-todo==1.6.2
fastcache==1.0.2
Pillow==4.1.1
ply==3.10
psutil==5.2.2
psycopg2==2.7.3.2
ptyprocess==0.5.1
pudb==2017.1.4
py==1.4.33
pycparser==2.17
pyOpenSSL==17.0.0
pyparsing==2.1.4
pystan==2.17.0.0
python-dateutil==2.6.0
python-memcached==1.58
pytz==2017.2
qtconsole==4.3.0
requests==2.14.2
simplegeneric==0.8.1
six==1.10.0
statsmodels==0.8.0
tables==3.3.0
tablib==0.12.1
tblib==1.3.2
XlsxWriter==0.9.6
